AFML_FINSIGHT/requirements.txt
google-generativeai
python-dotenv
pandas
yfinance
fredapi
sec-edgar-api
requests
plotly
kaleido
streamlit
